LayoutTests/http/tests/security/contentSecurityPolicy/eval-scripts-setTimeout-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to evaluate script because of Content-Security-Policy.

ALERT: PASS
